requirements.txt
autopep8==1.6.0
click==8.1.3
cycler==0.11.0
Flask==2.1.2
fonttools==4.33.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.22.4
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.1
pycodestyle==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
toml==0.10.2
Werkzeug==2.1.2
zipp==3.8.0
